requirements.txt
Pillow>=9.5.0
imagehash==4.3.1
python-dateutil==2.8.2
opencv-python==4.9.0.80
numpy==1.26.4
exifread==3.0.0
geopy==2.4.1